requirements.txt
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
PIMS==0.6.1
pytorch_lightning==1.6.2
pytorchvideo==0.1.5
PyYAML==6.0.1
ruptures==1.1.8
scikit_learn==1.0.2
scipy==1.13.0
termcolor==1.1.0
torch==1.8.1
torchbnn==1.2
torchvision==0.9.1
tqdm==4.51.0
tsai==0.3.6
torch-sgld==0.1.0
comet-ml==3.37.1
